Added Copy *without the facebook folder*
I followed the instructions, and got "390 duplicate symbol errors". 
I realized that I had included twice the FacebookSDK. Once in the Step 2 of "### Add the Facebook iOS and JavaScript SDK" and once in the step 1 of "### Add the Cordova Facebook Connect Plugin". Because it seems that in the "src/ios" folder, there is the code of the FacebookSDK.
Hope this help