Fix hasAccessToken check (#853)
When an access token is expired this plugin would not
fetch a new one. This is invalid behaviour and resulted
in failed login attempts because the token was expired,
but still passed to the javascript code as a valid token.

This patch adds a check if the token is expired as well.